Make player spawn one tile away from the base
