+++ Assets/Scenes/RandomMapScene.unity
@@ -200,6 +200,7 @@ MonoBehaviour:
   straightPrefab: {fileID: 7894887961081992580, guid: 4c85730c47d6f47ea907c33ec17975d2, type: 3}
   waypointPrefab: {fileID: 7330119712648875633, guid: 5c6966a955fed8f4a95b999a0d6ec8ab, type: 3}
   waypointEmpty: {fileID: 916810571}
+  playerGameObj: {fileID: 1556804005}
 --- !u!1 &138248409
 GameObject:
   m_ObjectHideFlags: 0
@@ -976,6 +977,7 @@ GameObject:
   - component: {fileID: 1556804006}
   - component: {fileID: 1556804010}
   - component: {fileID: 1556804009}
+  - component: {fileID: 1556804013}
   - component: {fileID: 1556804008}
   - component: {fileID: 1556804007}
   m_Layer: 10
@@ -1137,6 +1139,18 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1556804013
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1556804005}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4cf6ce614676e47adb869237d7be92f8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &1642223570
 GameObject:
   m_ObjectHideFlags: 0
