+++ Assets/Scripts/Player/PlayerShooting.cs
