+++ Assets/Scripts/Player/PlayerStartposition.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4cf6ce614676e47adb869237d7be92f8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
